Add first parts of markup pretty printer
The first part focusses on the bread and butter parts of the markup language:
the text plugin. This is a good starting point as this is the most important
editor plugin and the escape mechanisms which are needed to display the
contents correctly are good foundations for the other plugins.