Add data types for AsciiDoc header
This is only a first attempt of creating data types to represent the header of
an AsciiDoc file. This representation is not ideal: Everything stored in the
header is essentially meta data - only the syntax representing everything in
the file itself is different. Everything in the Header is optional (Except for
maybe the title if the header is present?). Thus instead of representing all
possible generally specified attributes in a generic algebraic data type, one
could use somthing like an association list or a hash map.