Add first features to paragraph parser
The paragraph parser understands nonspecial chars and hard line
breaks. Currently it creates a string parser instead of one of the types
specified in Text.AsciiDoc.Types.Text. Furhtermore there is a small bug, which
happens in an edge case when creating hard line breaks. The document
"Test +\n \nsomething" is parsed as "Test\n something" instead of
"Test\nsomething".