Add parser to parse entire document
Up to this point, there were only small parsers which could parse smaller
portions of the document like a paragraph or the document header. Combine
these to parse entire documents.

This combined parser has - of course - all the limitations of the individual
smaller parsers regarding edge cases and compliance with the spec.