Work around the article plugin
Serlo created a new Plugin called "article", which is the root plugin of
articles. Currently the converter is only interested in the article's
content. Hence the Article tags are stripped when the JSON is running through
fromJSON.

This solution is a quick fix and should be replaced by properly encoding the
article as a SerloPlugin.